Bug 662851 - Make EnsureWritableCapacity take uint32_t instead of uint64_t since all callers pass uint32_t (r=stejohns,sr-pending=ed)

+++ core/ByteArrayGlue.h
@@ -147,7 +147,7 @@ namespace avmplus
                 , m_oldLength(owner->m_length)
             {
             }
-            void FASTCALL EnsureWritableCapacity(uint64_t minimumCapacity);
+            void FASTCALL EnsureWritableCapacity(uint32_t minimumCapacity);
             ~Grower();
         
             ByteArray*  m_owner;
